requirements.txt
fastapi
uvicorn
anthropic
python-dotenv
httpx
stripe
